backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
pydantic-settings==2.9.1
pydantic==2.11.4
python-dotenv==1.1.0
openai==1.78.1
anthropic==0.51.0